// src/generated/models/Dyn365bc_productionorders_abk_prod_v1_0Model.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Dyn365bc_productionorders_abk_prod_v1_0dyn365bc_sourcetype = {
  200000000: 'Item',
  200000001: 'Family',
  200000002: 'SalesHeader'
} as const;
export type Dyn365bc_productionorders_abk_prod_v1_0dyn365bc_sourcetype = keyof typeof Dyn365bc_productionorders_abk_prod_v1_0dyn365bc_sourcetype;
export const Dyn365bc_productionorders_abk_prod_v1_0dyn365bc_status = {
  200000000: 'Simulated',
  200000001: 'Planned',
  200000002: 'FirmPlanned',
  200000003: 'Released',
  200000004: 'Finished'
} as const;
export type Dyn365bc_productionorders_abk_prod_v1_0dyn365bc_status = keyof typeof Dyn365bc_productionorders_abk_prod_v1_0dyn365bc_status;

export interface Dyn365bc_productionorders_abk_prod_v1_0Base {
  dyn365bc_assigneduserid?: string;
  dyn365bc_bincode?: string;
  dyn365bc_companylookup_code?: string;
  "dyn365bc_companylookup_id@odata.bind": string;
  dyn365bc_creationdate?: string;
  dyn365bc_description?: string;
  dyn365bc_duedate?: string;
  dyn365bc_endingdate?: string;
  dyn365bc_endingtime?: string;
  dyn365bc_finisheddate?: string;
  dyn365bc_genprodpostinggroup?: string;
  dyn365bc_inventorypostinggroup?: string;
  dyn365bc_lastdatemodified?: string;
  dyn365bc_locationcode?: string;
  dyn365bc_lowlevelcode?: number;
  dyn365bc_no?: string;
  dyn365bc_productionorder_abk_prod_v1_0id: string;
  dyn365bc_quantity?: number;
  dyn365bc_routingno?: string;
  dyn365bc_searchdescription?: string;
  dyn365bc_sourceno?: string;
  dyn365bc_sourcetype?: Dyn365bc_productionorders_abk_prod_v1_0dyn365bc_sourcetype;
  dyn365bc_startingdate?: string;
  dyn365bc_startingtime?: string;
  dyn365bc_status?: Dyn365bc_productionorders_abk_prod_v1_0dyn365bc_status;
}

export interface Dyn365bc_productionorders_abk_prod_v1_0 extends Dyn365bc_productionorders_abk_prod_v1_0Base {
  dyn365bc_companylookup_idname?: string;
  dyn365bc_sourcetypename?: string;
  dyn365bc_statusname?: string;
  dyn365bc_companylookup_id?: object;
  _dyn365bc_companylookup_id_value?: string;
}